ml-stat: remove stale "return" line
Commit cited via Fixes tag intended to add a second pattern to is_pr()
matching, but it left the previous logic intact. Fix it by removing
the stale return statement.
(I was adding to patch by one line, and somehow lost the removal).

Technically it is the moment that cited commit becomes effective.

Fixes: c58058fe939b ("ml-stat: add "pull-req" to PR patterns")
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>